requirements.txt
python-dotenv==1.0.0
streamlit==1.23.1
torch==2.0.1+cu118
transformers==4.29.2
streamlit-chat=0.0.2.2